# Conflicts:
#	src/main/java/za/ac/cput/domain/Supplier.java
#	src/main/java/za/ac/cput/repository/IOrderRepository.java
#	src/main/java/za/ac/cput/repository/ISupplierRepository.java
#	src/main/java/za/ac/cput/service/IOrderItemService.java
#	src/main/java/za/ac/cput/service/ISupplierService.java
#	src/test/java/za/ac/cput/domain/SupplierTest.java

+++ src/main/java/za/ac/cput/factory/EmployeeFactory.java
@@ -9,7 +9,7 @@ import za.ac.cput.domain.Employee;
 
 public class EmployeeFactory {
 
-    public static Employee createEmployee(int employeeID, String firstName, String lastName, String email, String position, double salary) {
+    public static Employee createEmployee(Long employeeID, String firstName, String lastName, String email, String position, double salary) {
         // Trim input values to avoid leading/trailing spaces being considered as valid inputs
         firstName = firstName != null ? firstName.trim() : "";
         lastName = lastName != null ? lastName.trim() : "";
